SysInfo to System_Stats
Changed the sysinfo script so it can be run via cron to write to motd
(like my old system_stats script… which has just been replaced). Adds
the auto command line argument to redirect output to motd; also
adjusted the out-Heading function to handle the lack of an output
device and “hard set” the length of the highlight.