Logname issues
Seems logname can’t pick the original user name if there’s no tty to
hook to. Some desktop envs or window managers or terminals or X systems
or ssh’s or… don’t establish a tty for logname to work. I try to catch
it and adjust as necessary.