Changed verbs; psSysInfo module; prompt change prep
I was not using approved verbs (lots of shows which should be gets). So
I changed a lot of them.

Added the call to the sysinfo module. Now I just need to upload that
module. (And can I just say, PS modules don't work quite the way they're
supposed to. From alias problems [no descriptions?!?! really?!?!] to
private functions that have to be manually set as private by naming them
wrong and exporting only right names!!)

I'm also laying the groundwork for separating the prompt out of the
profile into a module... assuming that will work....