Directories and Profile Cleanup
Just a little clean up on the directories module and the profile.
Directories got some display changes (date/time, color); profile removed
the alias for directories module (now in the directories module where it
should be) and removal of the credential functions. They're useful, but
not for me at this point.