Profile PSDrives w/Description; Fixed redundant Office
Adjusted the PSDrives created in the profile to include descriptions -
makes it easier to find (get-ProfilePSDrive) and understand them.

Adjusted the program alias function. A set of updates to Office created
an "Updates" folder with redundant (not fully fleshed out copies of
excel and winword) (that is, one didn't seem to have version info). So,
1) it "null errored" when trying to assign the 2nd alias (the one with
version info) and 2) bumped the "correct" program into a versioned alias
(word16 vs. word). I added a try/catch to prevent the error (such a
cheap fix - shoulda been there to start!) and a quick test/ignore for
Updates in the name.