Added Get-CurrentCalendar
Added a function to my profile that displays a quick calendar of the
current months + surrounding months. The output is very similar to a
bash script I use in Linux and MacOS, but the process is completely
different (windows lacks the cal program). There were a couple ways I
could've done this, but this was quick and simple.